Clean up OTP token request and AM generate token
