+++ src/types/one-time-password-token-request.d.ts
@@ -1,20 +1,20 @@
 /**
- * Onetime Password Token Request
+ * One Time Password Token Request
  * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/single-sign-on/user-authentication-password-profiles/index.html
  */
 
 /**
- * The Onetime Password Token Request object
+ * The One Time Password Token Request object
  * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/single-sign-on/user-authentication-password-profiles/index.html#The-User-Authentication-Password-Profile-Object
  */
 
-export interface OnetimePasswordTokenRequestBody {
+export interface OneTimePasswordTokenRequestBody {
     type: string
     username: string
     purpose : string
 }
 
-export interface OnetimePasswordTokenRequestEndpoint {
+export interface OneTimePasswordTokenRequestEndpoint {
     /**
      * Create a Onetime Password Token Request
      * DOCS: https://documentation.elasticpath.com/commerce-cloud/docs/api/single-sign-on/user-authentication-password-profiles/create-a-user-authentication-password-profile.html
@@ -22,6 +22,6 @@ export interface OnetimePasswordTokenRequestEndpoint {
      * @param passwordProfileId - The ID for the password profile.
      * @param body - The Onetime Password Token Request object
      */
-    Create(realmId: string, passwordProfileId: string, body: { data: OnetimePasswordTokenRequestBody }, token?: string)
+    Create(realmId: string, passwordProfileId: string, body: OneTimePasswordTokenRequestBody, token?: string)
 
 }
