+++ test/unit/one-time-password-token-request.ts
@@ -4,7 +4,7 @@ import { gateway as MoltinGateway } from '../../src/moltin'
 
 const apiUrl = 'https://euwest.api.elasticpath.com/v2'
 
-describe('User Authentication Password Profile Info', () => {
+describe('One Time Password Token Request', () => {
     const Moltin = MoltinGateway({
         client_id: 'XXX'
     })
@@ -14,16 +14,16 @@ describe('User Authentication Password Profile Info', () => {
 
     it('Create a One Time Password Token Info', () => {
         nock(apiUrl, {})
-            .post(/\/authentication-realms\/(.*)\/password-profiles\/(.*)\/onetime-password-request-token/)
-            .reply(201, {})
+            .post(/\/authentication-realms\/(.*)\/password-profiles\/(.*)\/one-time-password-token-request/)
+            .reply(202, {})
 
         const body = {
-            type: "one_time_password_request_token",
+            type: "one_time_password_token_request",
             username: "john.doe",
             purpose : "reset_password"
         }
 
-        return Moltin.OneTimePasswordTokenRequest.Create(realmId,passwordProfileId, { data: body }).then(res => {
+        return Moltin.OneTimePasswordTokenRequest.Create(realmId,passwordProfileId, body).then(res => {
             assert.isObject(res)
         })
     })
